refactor(installer): extract humanizer logic to InstallerHumanizer class
Fixes PHPCS violations: class too long, function too long, line too long.
Moves humanizer directive append logic to dedicated InstallerHumanizer class.
