+++ src/Installer.php
@@ -12,11 +12,6 @@ use SplFileInfo;
 final class Installer
 {
 
-    private const string HUMANIZER_SECTION_HEADING = '## Output Humanization';
-
-    private const string HUMANIZER_DIRECTIVE = '- Use [blader/humanizer](https://github.com/blader/humanizer)'
-        . ' for all skill outputs to keep the text natural and human-friendly.';
-
     /**
      * @param array<int, string> $argv
      */
@@ -109,9 +104,11 @@ final class Installer
             )
             : [0, 0];
 
-        $claudeMdCopied = self::installClaudeMd($root, $editor);
+        $claudeMdSource = InstallerPath::isClaudeMdEditor($editor) ? InstallerPath::resolveClaudeMdSource() : null;
+        $claudeMdCopied = self::installSingleFile($claudeMdSource, InstallerPath::resolveClaudeMdTarget($root));
+        $total = $rulesCopied + $skillsCopied + $claudeMdCopied;
 
-        echo sprintf('Cursor rules installed (%d files, %d pruned).%s', $rulesCopied + $skillsCopied + $claudeMdCopied, $rulesPruned + $skillsPruned, PHP_EOL);
+        echo sprintf('Cursor rules installed (%d files, %d pruned).%s', $total, $rulesPruned + $skillsPruned, PHP_EOL);
 
         return 0;
     }
@@ -230,42 +227,9 @@ final class Installer
             self::copy($src, $dst);
         }
 
-        self::appendHumanizerDirectiveIfNeeded($dst);
+        InstallerHumanizer::appendIfNeeded($dst);
 
         return true;
-    }
-
-    private static function appendHumanizerDirectiveIfNeeded(string $destination): void
-    {
-        if (!self::isSkillMarkdown($destination) || is_link($destination)) {
-            return;
-        }
-
-        $contents = file_get_contents($destination);
-
-        if ($contents === false || str_contains($contents, self::HUMANIZER_DIRECTIVE)) {
-            return;
-        }
-
-        $normalized = rtrim($contents);
-
-        if ($normalized !== '') {
-            $normalized .= PHP_EOL . PHP_EOL;
-        }
-
-        $updated = $normalized
-            . self::HUMANIZER_SECTION_HEADING . PHP_EOL
-            . self::HUMANIZER_DIRECTIVE
-            . PHP_EOL;
-
-        file_put_contents($destination, $updated);
-    }
-
-    private static function isSkillMarkdown(string $path): bool
-    {
-        $normalized = str_replace('\\', '/', $path);
-
-        return str_ends_with($normalized, '/SKILL.md');
     }
 
     private static function removeExistingTarget(string $destination): void
@@ -346,21 +310,9 @@ final class Installer
         }
     }
 
-    private static function installClaudeMd(string $root, string $editor): int
+    private static function installSingleFile(?string $source, string $target): int
     {
-        if (!InstallerPath::isClaudeMdEditor($editor)) {
-            return 0;
-        }
-
-        $source = InstallerPath::resolveClaudeMdSource();
-
-        if ($source === null) {
-            return 0;
-        }
-
-        $target = InstallerPath::resolveClaudeMdTarget($root);
-
-        if (file_exists($target)) {
+        if ($source === null || file_exists($target)) {
             return 0;
         }
